Move annotation object from main to annotation/annotation.
